okay first of all I created 2 AI one with  <AI prception>  and one with  <AI pawn Sencing> .
the secend one will patroll between two  "TargetPoint"  Actors and if see or hear you will followe you.

and the first one that was created with BT and BB will choose a rand location and go to that location and if see you also follow you.

the first one is in Content/AI/AIStuff1
the second one is in Content/AI/AIStuff2

unfortunatlly didm't mange to add a noise emitter to my projectile like my FPSCharctor (I mean I did but it did't worked)